Move serialization specific types to `serialization` subpackage
And move codecs to `serialization.codecs` subpackage.
